refactor : command
コマンド入力で攻撃が可能にした

予定　プレイヤーの動きを改善していく
主にダッシュ　回避　の動き

+++ MechaAction/Assets/MechaAction/Code/Enemy/BurstEnemy/Burst_Attack.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class Burst_Attack : MonoBehaviour
 {
-    private int m_damage;
+    private float m_damage;
     private float m_criticalRate;
     private float m_criticalChance;
-    private int m_knockback;
+    private float m_knockback;
     private AttackType m_type;
     private float m_duration;
